Hoist shared fractal UI domain assets
