+++ packages/shared-ui/domains/fractal/components/fractal-guild-header-identifier.tsx
