+++ packages/shared-ui/domains/fractal/lib/constants.ts
