+++ packages/shared-ui/domains/fractal/lib/getMemberLabel.ts
@@ -1,4 +1,4 @@
-import { MemberStatus } from "./zod/MemberStatus";
+import { MemberStatus } from "./schemas/MemberStatus";
 
 export const getMemberLabel = (status: MemberStatus): string =>
     status == MemberStatus.Citizen
