+++ packages/shared-ui/domains/fractal/lib/graphql/getFractal.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { FRACTALS_SERVICE } from "@/lib/constants";
+import { FRACTALS_SERVICE } from "../constants";
 
 import { graphql } from "@shared/lib/graphql";
 import { Account, zAccount } from "@shared/lib/schemas/account";
