+++ packages/user/FractalCore/ui/src/components/app-sidebar.tsx
@@ -7,6 +7,7 @@ import { useFractal } from "@/hooks/fractals/use-fractal";
 import { useFractalAccount } from "@/hooks/fractals/use-fractal-account";
 
 import { UserSidebarNavFooter } from "@shared/components/user-sidebar-nav-footer";
+import { FractalGuildIdentifier } from "@shared/domains/fractal/components/fractal-guild-header-identifier";
 import {
     Sidebar,
     SidebarContent,
@@ -17,8 +18,6 @@ import {
     SidebarMenuItem,
     SidebarRail,
 } from "@shared/shadcn/ui/sidebar";
-
-import { FractalGuildIdentifier } from "./fractal-guild-header-identifier";
 
 export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
     const navigate = useNavigate();
