+++ packages/user/FractalCore/ui/src/hooks/fractals/use-membership.ts
@@ -1,13 +1,13 @@
-import { queryClient } from "@shared/lib/queryClient";
 import { useQuery } from "@tanstack/react-query";
+
+import QueryKey, { OptionalAccount } from "@/lib/queryKeys";
 
 import {
     Membership,
     getMembership,
-} from "@/lib/graphql/fractals/getMembership";
-import QueryKey, { OptionalAccount } from "@/lib/queryKeys";
-import { MemberStatus } from "@/lib/zod/MemberStatus";
-
+} from "@shared/domains/fractal/lib/graphql/getMembership";
+import { MemberStatus } from "@shared/domains/fractal/lib/schemas/MemberStatus";
+import { queryClient } from "@shared/lib/queryClient";
 import { zAccount } from "@shared/lib/schemas/account";
 
 const queryFn = async (fractal: string, user: string) => {
