+++ packages/user/FractalCore/ui/src/hooks/fractals/use-users-and-groups.ts
@@ -3,13 +3,13 @@ import type { Account } from "@shared/lib/schemas/account";
 import { useQuery } from "@tanstack/react-query";
 import { z } from "zod";
 
-import { FRACTALS_SERVICE } from "@/lib/constants";
 import {
     getUsersAndGroups,
     zUsersAndGroupsResponse,
 } from "@/lib/graphql/evaluations/getUsersAndGroups";
 import QueryKey from "@/lib/queryKeys";
 
+import { FRACTALS_SERVICE } from "@shared/domains/fractal/lib/constants";
 import { assertUser } from "@shared/hooks/use-current-user";
 import { queryClient } from "@shared/lib/queryClient";
 import { updateArray } from "@shared/lib/update-array";
