+++ packages/user/FractalCore/ui/src/lib/graphql/evaluations/getEvaluation.ts
@@ -1,8 +1,11 @@
 import { z } from "zod";
 
-import { EVALUATIONS_SERVICE, FRACTALS_SERVICE } from "@/lib/constants";
 import { zUnix } from "@/lib/zod/Unix";
 
+import {
+    EVALUATIONS_SERVICE,
+    FRACTALS_SERVICE,
+} from "@shared/domains/fractal/lib/constants";
 import { graphql } from "@shared/lib/graphql";
 import { zAccount } from "@shared/lib/schemas/account";
