+++ packages/user/FractalCore/ui/src/lib/graphql/evaluations/getUsersAndGroups.ts
@@ -1,7 +1,6 @@
 import { z } from "zod";
 
-import { EVALUATIONS_SERVICE } from "@/lib/constants";
-
+import { EVALUATIONS_SERVICE } from "@shared/domains/fractal/lib/constants";
 import { graphql } from "@shared/lib/graphql";
 import { Account, zAccount } from "@shared/lib/schemas/account";
