+++ packages/user/FractalCore/ui/src/lib/graphql/fractals/getCompletedEvaluations.ts
@@ -1,8 +1,7 @@
 import dayjs from "dayjs";
 import { z } from "zod";
 
-import { FRACTALS_SERVICE } from "@/lib/constants";
-
+import { FRACTALS_SERVICE } from "@shared/domains/fractal/lib/constants";
 import { graphql } from "@shared/lib/graphql";
 import { Account, zAccount } from "@shared/lib/schemas/account";
