+++ packages/user/FractalCore/ui/src/lib/graphql/fractals/getGuildApplication.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
-import { FRACTALS_SERVICE } from "@/lib/constants";
 import { zGuildApplicationListInstance } from "@/lib/zod/attestations";
 
+import { FRACTALS_SERVICE } from "@shared/domains/fractal/lib/constants";
 import { graphql } from "@shared/lib/graphql";
 import { Account } from "@shared/lib/schemas/account";
