+++ packages/user/FractalCore/ui/src/lib/graphql/fractals/getMembers.ts
@@ -1,8 +1,7 @@
 import { z } from "zod";
 
-import { FRACTALS_SERVICE } from "@/lib/constants";
-import { MemberStatus } from "@/lib/zod/MemberStatus";
-
+import { FRACTALS_SERVICE } from "@shared/domains/fractal/lib/constants";
+import { MemberStatus } from "@shared/domains/fractal/lib/schemas/MemberStatus";
 import { graphql } from "@shared/lib/graphql";
 import { Account, zAccount } from "@shared/lib/schemas/account";
 import { zDateTime } from "@shared/lib/schemas/date-time";
