+++ packages/user/FractalCore/ui/src/pages/fractal/members.tsx
@@ -5,11 +5,11 @@ import { PageContainer } from "@/components/page-container";
 
 import { useFractalAccount } from "@/hooks/fractals/use-fractal-account";
 import { useMembers } from "@/hooks/fractals/use-members";
-import { COUNCIL_SEATS } from "@/lib/constants";
-import { getMemberLabel } from "@/lib/getMemberLabel";
 
 import { GlowingCard } from "@shared/components/glowing-card";
 import { TableContact } from "@shared/components/tables/table-contact";
+import { COUNCIL_SEATS } from "@shared/domains/fractal/lib/constants";
+import { getMemberLabel } from "@shared/domains/fractal/lib/getMemberLabel";
 import { Badge } from "@shared/shadcn/ui/badge";
 import { CardContent, CardHeader, CardTitle } from "@shared/shadcn/ui/card";
 import {
