+++ packages/user/Fractals/ui/src/components/fractal-switcher.tsx
@@ -8,9 +8,9 @@ import { useMemberships } from "@/hooks/fractals/use-memberships";
 import { useCurrentFractal } from "@/hooks/use-current-fractal";
 import { useCurrentUser } from "@/hooks/use-current-user";
 import { createIdenticon } from "@/lib/createIdenticon";
-import { zFractal } from "@/lib/graphql/fractals/getFractal";
 import QueryKey from "@/lib/queryKeys";
 
+import { zFractal } from "@shared/domains/fractal/lib/graphql/getFractal";
 import { useChainId } from "@shared/hooks/use-chain-id";
 import {
     DropdownMenu,
