+++ packages/user/Fractals/ui/src/hooks/fractals/use-membership.ts
@@ -1,13 +1,14 @@
 import { queryClient } from "@/queryClient";
 import { useQuery } from "@tanstack/react-query";
 
-import {
-    Membership,
-    getMembership,
-} from "@/lib/graphql/fractals/getMembership";
 import QueryKey, { OptionalAccount } from "@/lib/queryKeys";
-import { zAccount } from "@/lib/zod/Account";
-import { MemberStatus } from "@/lib/zod/MemberStatus";
+
+import {
+    type Membership,
+    getMembership,
+} from "@shared/domains/fractal/lib/graphql/getMembership";
+import { MemberStatus } from "@shared/domains/fractal/lib/schemas/MemberStatus";
+import { zAccount } from "@shared/lib/schemas/account";
 
 const queryFn = async (fractal: string, user: string) => {
     try {
