+++ packages/user/Fractals/ui/src/hooks/use-account-status.ts
@@ -4,7 +4,8 @@ import { z } from "zod";
 import { supervisor } from "@/supervisor";
 
 import QueryKey from "@/lib/queryKeys";
-import { zAccount } from "@/lib/zod/Account";
+
+import { zAccount } from "@shared/lib/schemas/account";
 
 export const zAccountNameStatus = z.enum(["Available", "Taken", "Invalid"]);
 const zGetAccountReturn = z
