+++ packages/user/Fractals/ui/src/hooks/use-current-user.ts
@@ -5,7 +5,8 @@ import { z } from "zod";
 import { supervisor } from "@/supervisor";
 
 import QueryKey from "@/lib/queryKeys";
-import { Account, zAccount } from "@/lib/zod/Account";
+
+import { type Account, zAccount } from "@shared/lib/schemas/account";
 
 export type GetCurrentUserRes = string | null;
