+++ packages/user/Fractals/ui/src/hooks/use-select-account.ts
@@ -4,8 +4,8 @@ import { useMutation } from "@tanstack/react-query";
 import { supervisor } from "@/supervisor";
 
 import QueryKey from "@/lib/queryKeys";
-import { zAccount } from "@/lib/zod/Account";
 
+import { zAccount } from "@shared/lib/schemas/account";
 import { toast } from "@shared/shadcn/ui/sonner";
 
 export const useSelectAccount = () =>
