--- packages/user/Fractals/ui/src/lib/constants.ts
@@ -1,4 +1,0 @@
-import { zAccount } from "./zod/Account";
-
-export const fractalsService = zAccount.parse("fractals");
-export const evaluationsService = zAccount.parse("evaluations");
