--- packages/user/Fractals/ui/src/lib/getMemberLabel.ts
@@ -1,8 +1,0 @@
-import { MemberStatus } from "./zod/MemberStatus";
-
-export const getMemberLabel = (status: MemberStatus): string =>
-    status == MemberStatus.Citizen
-        ? "Citizen"
-        : status == MemberStatus.Visa
-          ? "Visa"
-          : "Exiled";
