+++ packages/user/Fractals/ui/src/lib/graphql/evaluations/getEvaluation.ts
@@ -2,10 +2,11 @@ import { z } from "zod";
 
 import { siblingUrl } from "@psibase/common-lib";
 
-import { fractalsService } from "@/lib/constants";
 import { graphql } from "@/lib/graphql";
-import { zAccount } from "@/lib/zod/Account";
 import { zUnix } from "@/lib/zod/Unix";
+
+import { FRACTALS_SERVICE } from "@shared/domains/fractal/lib/constants";
+import { zAccount } from "@shared/lib/schemas/account";
 
 export const zEvaluation = z.object({
     id: z.number(),
@@ -26,7 +27,7 @@ export const getEvaluation = async (id: number) => {
     const evaluation = await graphql(
         `
     {
-        getEvaluation(owner: "${fractalsService}", evaluationId: ${id}) {     
+        getEvaluation(owner: "${FRACTALS_SERVICE}", evaluationId: ${id}) {     
             id,
             createdAt,
             owner,
