+++ packages/user/Fractals/ui/src/lib/graphql/fractals/getFractals.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 import { siblingUrl } from "@psibase/common-lib";
 
-import { zAccount } from "@/lib/zod/Account";
+import { zAccount } from "@shared/lib/schemas/account";
 
 import { graphql } from "../../graphql";
