+++ packages/user/Fractals/ui/src/lib/graphql/fractals/getMembers.ts
@@ -2,9 +2,10 @@ import { z } from "zod";
 
 import { siblingUrl } from "@psibase/common-lib";
 
-import { Account, zAccount } from "@/lib/zod/Account";
 import { zDateTime } from "@/lib/zod/DateTime";
-import { MemberStatus } from "@/lib/zod/MemberStatus";
+
+import { MemberStatus } from "@shared/domains/fractal/lib/schemas/MemberStatus";
+import { type Account, zAccount } from "@shared/lib/schemas/account";
 
 import { graphql } from "../../graphql";
