--- packages/user/Fractals/ui/src/lib/zod/MemberStatus.ts
@@ -1,9 +1,0 @@
-import { z } from "zod";
-
-export enum MemberStatus {
-    Visa = 1,
-    Citizen = 2,
-    Exiled = 3,
-}
-
-export const zMemberStatus = z.nativeEnum(MemberStatus);
