+++ packages/user/Fractals/ui/src/pages/fractal/membership/my-membership.tsx
@@ -1,5 +1,5 @@
-import type { FractalRes } from "@/lib/graphql/fractals/getFractal";
-import type { Membership } from "@/lib/graphql/fractals/getMembership";
+import type { FractalRes } from "@shared/domains/fractal/lib/graphql/getFractal";
+import type { Membership } from "@shared/domains/fractal/lib/graphql/getMembership";
 
 import dayjs from "dayjs";
 
@@ -13,9 +13,9 @@ import { useMembership } from "@/hooks/fractals/use-membership";
 import { useCurrentFractal } from "@/hooks/use-current-fractal";
 import { useCurrentUser } from "@/hooks/use-current-user";
 import { createIdenticon } from "@/lib/createIdenticon";
-import { getMemberLabel } from "@/lib/getMemberLabel";
 
 import { GlowingCard } from "@shared/components/glowing-card";
+import { getMemberLabel } from "@shared/domains/fractal/lib/getMemberLabel";
 import { useChainId } from "@shared/hooks/use-chain-id";
 import { Badge } from "@shared/shadcn/ui/badge";
 import { Button } from "@shared/shadcn/ui/button";
